app/requirements/postgres.txt
django-extensions==3.0.9
Django==3.2.15
django-bootstrap-modal-forms==2.0.0
django-crispy-forms==1.9.2
django-jsonview==2.0.0
django-sass_processor
django-q==1.3.4
gunicorn==20.0.4
numpy>=1.22.4
openpyxl
httpx==0.18.1
psycopg2-binary==2.9.3
XlsxWriter==1.3.9
requests==2.24.0
exchangelib==4.4.0
jsonschema==4.4.0
oemof-thermal==0.0.5
asgiref==3.6.0
blinker==1.5
channels==4.0.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.8.1
dash-bootstrap-components==1.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dill==0.3.6
django-plotly-dash==2.1.3
dpd-components==0.1.0
Flask==2.2.2
fonttools==4.38.0
graphviz==0.20.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.3
networkx==3.0
nose==1.3.7
numpy==1.23.5
git+https://github.com/oemof/oemof_visio.git
oemof.network==0.4.0
oemof.solph==0.4.5
oemof.tools==0.4.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.0
ply==3.11
Pyomo==5.7.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyUtilib==6.0.0
six==1.16.0
sqlparse==0.4.3
tenacity==8.1.0
Werkzeug==2.2.2
uvicorn